+++ rotkehlchen/chain/ethereum/modules/giveth/constants.py
@@ -1,0 +1,5 @@
+from typing import Final
+
+from rotkehlchen.chain.evm.types import string_to_evm_address
+
+GIVETH_DONATION_CONTRACT_ADDRESS: Final = string_to_evm_address('0x97b2cb568e0880B99Cd16EFc6edFF5272Aa02676')  # noqa: E501