+++ rotkehlchen/chain/optimism/modules/giveth/constants.py
@@ -4,3 +4,4 @@ from rotkehlchen.chain.evm.types import string_to_evm_address
 
 GIVPOW_ADDRESS: Final = string_to_evm_address('0x301C739CF6bfb6B47A74878BdEB13f92F13Ae5E7')
 GIV_TOKEN_ID: Final = 'eip155:10/erc20:0x528CDc92eAB044E1E39FE43B9514bfdAB4412B98'
+GIVETH_DONATION_CONTRACT_ADDRESS: Final = string_to_evm_address('0x8D685A56C51Cf54685d3dB0Ea50748D3A2c2e0dC')  # noqa: E501